+++ trading_strategy_tester/models/trading_result.py
@@ -2,6 +2,7 @@
 стратегии на конкретную дату. """
 
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -11,24 +12,24 @@ class TradingResult:
 
     Атрибуты:
         date_str (str): Дата торгов в формате 'YYYY-MM-DD'.
-        max_price (float): Максимальная цена за день.
-        min_price (float): Минимальная цена за день.
-        cache (float): Остаток денежных средств на конец дня.
+        max_price (Decimal): Максимальная цена за день.
+        min_price (Decimal): Минимальная цена за день.
+        cache (Decimal): Остаток денежных средств на конец дня.
         share_count (int): Количество акций в портфеле на конец дня.
-        amount_in_shares (float): Стоимость акций в портфеле на конец дня.
-        overall_result (float): Общая стоимость портфеля на конец дня.
-        comiss_sum (float): Сумма комиссии, уплаченной за день.
-        tax_sum (float): Сумма налога, уплаченного за день.
-        total_tax (float): Общая сумма налога, уплаченного за весь период.
+        amount_in_shares (Decimal): Стоимость акций в портфеле на конец дня.
+        overall_result (Decimal): Общая стоимость портфеля на конец дня.
+        comiss_sum (Decimal): Сумма комиссии, уплаченной за день.
+        tax_sum (Decimal): Сумма налога, уплаченного за день.
+        total_tax (Decimal): Общая сумма налога, уплаченного за весь период.
     """
 
     date_str: str
-    max_price: float
-    min_price: float
-    cache: float
+    max_price: Decimal
+    min_price: Decimal
+    cache: Decimal
     share_count: int
-    amount_in_shares: float
-    overall_result: float
-    comiss_sum: float
-    tax_sum: float
-    total_tax: float
+    amount_in_shares: Decimal
+    overall_result: Decimal
+    comiss_sum: Decimal
+    tax_sum: Decimal
+    total_tax: Decimal
